+++ packages/thumbprint-tokens/src/index.js
@@ -2,12 +2,34 @@
 const path = require('path');
 const glob = require('glob');
 const fse = require('fs-extra');
+const JSZip = require('jszip');
 const handlebars = require('handlebars');
 
 const outputs = [
     { slug: 'javascript-cjs', distName: 'index.js' },
     { slug: 'javascript-es', distName: 'index.js' },
     { slug: 'scss', distName: '_index.scss' },
+    {
+        slug: 'ios',
+        distName: 'ThumbprintTokens.swift',
+        // Packages the iOS dist into a zip file.
+        postWrite: async distPath => {
+            // Prep the zip file.
+            const zip = new JSZip();
+
+            // Add files to the folder.
+            zip.file('ThumbprintTokens.swift', fse.readFileSync(distPath, 'utf-8'));
+            zip.file('LICENSE.txt', fse.readFileSync('../../LICENSE', 'utf-8'));
+
+            // Save the zip file.
+            const encodedZip = await zip.generateAsync({ type: 'nodebuffer' });
+            const dist = path.resolve(__dirname, `../dist/ios.zip`);
+            await fse.outputFile(dist, encodedZip);
+
+            // We delete the `dist/ios` folder because we only want to keep the `.zip`.
+            await fse.remove('dist/ios');
+        },
+    },
 ];
 
 const compile = (output, tokens) => {
@@ -32,9 +54,15 @@ const compile = (output, tokens) => {
     // All of the files and tokens as a big array.
     const allTokens = allFiles.map(fileText => JSON.parse(fileText));
 
-    outputs.forEach(({ slug, distName }) => {
+    outputs.forEach(async ({ slug, distName, postWrite }) => {
         const contents = compile(slug, allTokens);
         const dist = path.resolve(__dirname, `../dist/${slug}/${distName}`);
-        fse.outputFileSync(dist, contents);
+
+        await fse.outputFile(dist, contents);
+
+        // Provide a hook for platform-specific processing at the end.
+        if (postWrite) {
+            await postWrite(dist);
+        }
     });
 })();