+++ packages/thumbprint-tokens/src/index.test.js
@@ -1,3 +1,4 @@
+const JSZip = require('jszip');
 const fs = require('fs');
 
 it('outputted CommonJS matches snapshot', () => {
@@ -18,5 +19,15 @@ it('outputted ES module matches snapshot', () => {
 
 it('outputted SCSS matches snapshot', () => {
     const output = fs.readFileSync('packages/thumbprint-tokens/dist/scss/_index.scss', 'utf-8');
+    expect(output).toMatchSnapshot();
+});
+
+it('outputted iOS matches snapshot', async () => {
+    // The output is stored as a `zip`, but we unzip it for the purpose of this test so that we
+    // can compare the token file as a string.
+    const zipBuffer = fs.readFileSync('packages/thumbprint-tokens/dist/ios.zip');
+    const files = await JSZip.loadAsync(zipBuffer);
+    const output = await files.file('ThumbprintTokens.swift').async('string');
+
     expect(output).toMatchSnapshot();
 });